+++ examples/resources/meraki_wireless_location_scanning/resource.tf
@@ -1,0 +1,5 @@
+resource "meraki_wireless_location_scanning" "example" {
+  network_id  = "L_123456"
+  enabled     = true
+  api_enabled = false
+}